Separate jobs cache functions
